remove AbstractRuntimeContext
