+++ sharding-jdbc/sharding-jdbc-core/src/main/java/org/apache/shardingsphere/shardingjdbc/jdbc/core/context/impl/ShardingRuntimeContext.java
@@ -19,24 +19,52 @@ package org.apache.shardingsphere.shardingjdbc.jdbc.core.context.impl;
 
 import com.google.common.collect.ImmutableMap;
 import lombok.Getter;
-import org.apache.shardingsphere.shardingjdbc.jdbc.core.context.AbstractRuntimeContext;
+import lombok.Setter;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.shardingsphere.core.log.ConfigurationLogger;
+import org.apache.shardingsphere.shardingjdbc.jdbc.core.context.RuntimeContext;
 import org.apache.shardingsphere.shardingjdbc.jdbc.core.datasource.metadata.CachedDatabaseMetaData;
+import org.apache.shardingsphere.sql.parser.SQLParserEngine;
+import org.apache.shardingsphere.sql.parser.SQLParserEngineFactory;
 import org.apache.shardingsphere.transaction.ShardingTransactionManagerEngine;
+import org.apache.shardingsphere.underlying.common.config.DatabaseAccessConfiguration;
+import org.apache.shardingsphere.underlying.common.config.properties.ConfigurationProperties;
+import org.apache.shardingsphere.underlying.common.config.properties.ConfigurationPropertyKey;
 import org.apache.shardingsphere.underlying.common.database.type.DatabaseType;
+import org.apache.shardingsphere.underlying.common.database.type.DatabaseTypes;
+import org.apache.shardingsphere.underlying.common.metadata.ShardingSphereMetaData;
+import org.apache.shardingsphere.underlying.common.metadata.datasource.DataSourceMetas;
+import org.apache.shardingsphere.underlying.common.metadata.schema.RuleSchemaMetaData;
+import org.apache.shardingsphere.underlying.common.metadata.schema.RuleSchemaMetaDataLoader;
 import org.apache.shardingsphere.underlying.common.rule.BaseRule;
+import org.apache.shardingsphere.underlying.executor.kernel.ExecutorKernel;
 
 import javax.sql.DataSource;
 import java.sql.Connection;
+import java.sql.DatabaseMetaData;
 import java.sql.SQLException;
 import java.util.Collection;
+import java.util.LinkedHashMap;
 import java.util.Map;
+import java.util.Map.Entry;
 import java.util.Properties;
 
 /**
- * Runtime context for sharding.
+ * Sharding runtime context.
  */
 @Getter
-public final class ShardingRuntimeContext extends AbstractRuntimeContext {
+@Slf4j(topic = "ShardingSphere-metadata")
+public final class ShardingRuntimeContext implements RuntimeContext {
+    
+    private final Collection<BaseRule> rules;
+    
+    private final ConfigurationProperties properties;
+    
+    private final DatabaseType databaseType;
+    
+    private final ExecutorKernel executorKernel;
+    
+    private final SQLParserEngine sqlParserEngine;
     
     private final CachedDatabaseMetaData cachedDatabaseMetaData;
     
@@ -44,12 +72,23 @@ public final class ShardingRuntimeContext extends AbstractRuntimeContext {
     
     private final Map<String, DataSource> dataSourceMap;
     
+    @Setter
+    private ShardingSphereMetaData metaData;
+    
     public ShardingRuntimeContext(final Map<String, DataSource> dataSourceMap, final Collection<BaseRule> rules, final Properties props, final DatabaseType databaseType) throws SQLException {
-        super(dataSourceMap, rules, props, databaseType);
+        this.rules = rules;
+        properties = new ConfigurationProperties(null == props ? new Properties() : props);
+        this.databaseType = databaseType;
+        executorKernel = new ExecutorKernel(properties.<Integer>getValue(ConfigurationPropertyKey.EXECUTOR_SIZE));
+        sqlParserEngine = SQLParserEngineFactory.getSQLParserEngine(DatabaseTypes.getTrunkDatabaseTypeName(databaseType));
         this.dataSourceMap = dataSourceMap;
         cachedDatabaseMetaData = createCachedDatabaseMetaData(dataSourceMap);
         shardingTransactionManagerEngine = new ShardingTransactionManagerEngine();
         shardingTransactionManagerEngine.init(databaseType, dataSourceMap);
+        metaData = createMetaData(dataSourceMap, databaseType);
+        // TODO log multiple rules
+        ConfigurationLogger.log(rules.iterator().next().getRuleConfiguration());
+        ConfigurationLogger.log(props);
     }
     
     public ShardingRuntimeContext(final DataSource dataSource, final Collection<BaseRule> rules, final Properties props, final DatabaseType databaseType) throws SQLException {
@@ -62,9 +101,30 @@ public final class ShardingRuntimeContext extends AbstractRuntimeContext {
         }
     }
     
+    private ShardingSphereMetaData createMetaData(final Map<String, DataSource> dataSourceMap, final DatabaseType databaseType) throws SQLException {
+        long start = System.currentTimeMillis();
+        DataSourceMetas dataSourceMetas = new DataSourceMetas(databaseType, getDatabaseAccessConfigurationMap(dataSourceMap));
+        RuleSchemaMetaData ruleSchemaMetaData = new RuleSchemaMetaDataLoader(rules).load(getDatabaseType(), dataSourceMap, getProperties());
+        ShardingSphereMetaData result = new ShardingSphereMetaData(dataSourceMetas, ruleSchemaMetaData);
+        log.info("Meta data load finished, cost {} milliseconds.", System.currentTimeMillis() - start);
+        return result;
+    }
+    
+    private Map<String, DatabaseAccessConfiguration> getDatabaseAccessConfigurationMap(final Map<String, DataSource> dataSourceMap) throws SQLException {
+        Map<String, DatabaseAccessConfiguration> result = new LinkedHashMap<>(dataSourceMap.size(), 1);
+        for (Entry<String, DataSource> entry : dataSourceMap.entrySet()) {
+            DataSource dataSource = entry.getValue();
+            try (Connection connection = dataSource.getConnection()) {
+                DatabaseMetaData metaData = connection.getMetaData();
+                result.put(entry.getKey(), new DatabaseAccessConfiguration(metaData.getURL(), metaData.getUserName(), null));
+            }
+        }
+        return result;
+    }
+    
     @Override
     public void close() throws Exception {
         shardingTransactionManagerEngine.close();
-        super.close();
+        executorKernel.close();
     }
 }
